requirements.txt
absl-py==2.4.0
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
chex==0.1.91
click==8.2.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
datasets==4.4.2
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
dm-acme==0.2.4
dm-env==1.6
dm-sonnet==2.0.2
dm-tree==0.1.9
evaluate==0.4.6
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.128.1
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.58.0
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.78.0
gymnasium==1.2.3
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jax==0.9.2
jaxlib==0.9.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.13.2
kiwisolver==1.4.8
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
namex==0.1.0
narwhals==1.40.0
nest-asyncio==1.6.0
networkx==3.6
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.19.0
packaging==24.2
pandas==2.3.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
polars==1.36.1
polars-runtime-32==1.36.1
prompt_toolkit==3.0.51
propcache==0.4.1
protobuf==7.34.1
psutil==7.0.0
pure_eval==0.2.3
pyaml==26.2.1
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.3
pyzmq==26.4.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.3
rich==14.3.3
rpds-py==0.25.1
safetensors==0.7.0
scikit-learn==1.8.0
scikit-optimize==0.10.2
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
stack-data==0.6.3
starlette==0.50.0
streamlit==1.45.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorflow==2.21.0
tensorflow-probability==0.25.0
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
toolz==1.1.0
torch==2.10.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trfl==1.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.235
ultralytics-thop==2.0.18
urllib3==2.4.0
uvicorn==0.40.0
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.46.3
wrapt==2.1.2
xxhash==3.6.0
yarl==1.22.0
